backend/requirements.txt
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
Flask==3.1.2
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.3.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msal==1.34.0
pika==1.3.2
pycparser==3.0
PyJWT==2.10.1
pyodbc==5.3.0
python-dotenv==1.2.1
requests==2.32.5
SQLAlchemy==2.0.46
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
azure-keyvault-secrets==4.10.0
azure-identity==1.24.0
